Fix ec2uploadimg when large helper instance used
This patch resolves an issue where a seconday disk is attached
to larger instance sizes and a seconday disk size is non integer.
For example, 139.7G. An Exception has been added to ignore this
error and print a message about the disk being skipped.

This closes https://github.com/SUSE-Enceladus/ec2imgutils/issues/46

+++ lib/ec2imgutils/ec2uploadimg.py
@@ -647,7 +647,7 @@ class EC2ImageUploader(EC2ImgUtils):
                 continue
             this_device_size = device['size']
             unit = this_device_size[-1]
-            size = int(this_device_size[:-1])
+            size = float(this_device_size[:-1])
             size_multiplier = 1
             if unit == 'T':
                 size_multiplier = 1024
